Add delete functionality for organization

+++ backend/src/modules/organization/repository/delete.repository.ts
@@ -1,4 +1,4 @@
-import { doc, deleteDoc } from "firebase/firestore";
+import { doc, getDoc, deleteDoc } from "firebase/firestore";
 import db from "../../../database/firebase.database";
 import { CustomError } from "../../../utils/customError";
 
@@ -11,9 +11,20 @@ export default async function organizationDelete({
 }) {
   try {
     const docRef = doc(db, "users", userId, "organizations", orgId);
-    await deleteDoc(docRef);
+
+    const docSnap = await getDoc(docRef);
+
+    if (!docSnap.exists()) {
+      throw new CustomError("NOT_FOUND", "Organization not found");
+    } else {
+      await deleteDoc(docRef);
+    }
   } catch (error: any) {
     console.error("#4678321579 Error deleting organization", error);
-    throw new CustomError("DB_ERROR", "Error deleting organization");
+    if (error.errorCode === "NOT_FOUND") {
+      throw new CustomError(error.errorCode, error.message);
+    } else {
+      throw new CustomError("DB_ERROR", "Error deleting organization");
+    }
   }
 }
